feat: configure storybook to work with Dashkit styles and palywright test-runner
