+++ .storybook/main.ts
@@ -1,12 +1,13 @@
 import type { StorybookConfig } from '@storybook/experimental-nextjs-vite';
 
 const config: StorybookConfig = {
-    stories: ['../app/**/*.mdx', '../app/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
+    stories: ['../app/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
     addons: ['@storybook/addon-essentials', '@storybook/experimental-addon-test'],
     framework: {
         name: '@storybook/experimental-nextjs-vite',
         options: {},
     },
+
     staticDirs: ['../public'],
 };
 export default config;
