+++ .storybook/manager.ts
@@ -1,0 +1,6 @@
+import { addons } from '@storybook/manager-api';
+import { themes } from '@storybook/theming';
+
+addons.setConfig({
+    theme: themes.light,
+});
