+++ app/components/account/TokenExtensionsCard.tsx
@@ -8,8 +8,9 @@ import { Cluster } from '@utils/cluster';
 import { useMemo } from 'react';
 import useSWR from 'swr';
 
+import { populatePartialParsedTokenExtension } from '@/app/utils/token-extension';
 import { getTokenInfo, getTokenInfoSwrKey } from '@/app/utils/token-info';
-import { TokenExtension, TokenExtensionType } from '@/app/validators/accounts/token-extension';
+import { TokenExtension } from '@/app/validators/accounts/token-extension';
 
 import { LoadingCard } from '../common/LoadingCard';
 import { TokenExtensionsSection } from './TokenExtensionsSection';
@@ -68,257 +69,4 @@ function populateTokenExtensions(extensions: TokenExtension[]): ParsedTokenExten
     }, new Map<string, ParsedTokenExtension>());
 
     return Array.from(result.values());
-}
-
-function populateSolanaDevelopersLink(component: string) {
-    return `https://solana.com/developers/guides/token-extensions/${component}`;
-}
-
-function populatePartialParsedTokenExtension(
-    extension: TokenExtensionType
-): Omit<ParsedTokenExtension, 'parsed' | 'extension'> {
-    function populateExternalLinks(url: string) {
-        return [{ label: 'Docs', url }];
-    }
-
-    switch (extension) {
-        case 'transferFeeAmount': {
-            const description =
-                "Every transfer sets aside a fee in the recipient's Token Account that can only be withdrawn by the Withdraw Authority";
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('transfer-fee')),
-                name: 'Transfer Fee Amount',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'mintCloseAuthority': {
-            const description = 'Allows a designated Close Authority to close the mint account if the supply is 0';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('mint-close-authority')),
-                name: 'Mint Close Authority',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'defaultAccountState': {
-            const description = 'Enables the authority to make new token accounts as frozen by default upon creation';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('default-account-state')),
-                name: 'Default Account State',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'immutableOwner': {
-            const description = 'Prevents the owner from being changed';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('immutable-owner')),
-                name: 'Immutable Owner',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'memoTransfer': {
-            const description = 'Requires all incoming transfers to a token account include a memo';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('required-memo')),
-                name: 'Required Memo',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'nonTransferable': {
-            return {
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('non-transferable-token')),
-                name: 'Non-Transferable Token',
-                status: 'active',
-            };
-            break;
-        }
-        case 'nonTransferableAccount': {
-            return {
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('non-transferable-token')),
-                name: 'Non-Transferable Token Account',
-                status: 'active',
-            };
-            break;
-        }
-        case 'cpiGuard': {
-            const description = 'Prohibits certain actions inside cross-program invocations';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('cpi-guard')),
-                name: 'CPI Guard',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'permanentDelegate': {
-            const description =
-                'Delegates permanent authority to a specific address that can transfer or burn tokens from any account holding this token';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('permanent-delegate')),
-                name: 'Permanent Delegate',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'transferHook': {
-            const description = 'Allow the token program to execute custom instruction logic on every token transfer';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('transfer-hook')),
-                name: 'Transfer Hook',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'transferHookAccount': {
-            return {
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('transfer-hook')),
-                name: 'Transfer Hook Account Info',
-                status: 'active',
-            };
-            break;
-        }
-        case 'metadataPointer': {
-            const description = 'Describes the location of the token metadata';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('metadata-pointer')),
-                name: 'Metadata Pointer',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'groupPointer': {
-            return {
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('group-member')),
-                name: 'Group Pointer',
-                status: 'active',
-            };
-            break;
-        }
-        case 'groupMemberPointer': {
-            return {
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('group-member')),
-                name: 'Group Member Pointer',
-                status: 'active',
-            };
-            break;
-        }
-        case 'confidentialTransferAccount': {
-            const description = 'Token amount is only able to be unencrypted by the token holder or auditor key';
-            return {
-                description,
-                externalLinks: populateExternalLinks('https://spl.solana.com/confidential-token/quickstart'),
-                name: 'Confidential Transfer Token Info',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'confidentialTransferFeeConfig': {
-            return {
-                externalLinks: populateExternalLinks('https://spl.solana.com/confidential-token/quickstart'),
-                name: 'Confidential Transfer Fee Config',
-                status: 'active',
-            };
-            break;
-        }
-        case 'confidentialTransferFeeAmount': {
-            return {
-                externalLinks: populateExternalLinks('https://spl.solana.com/confidential-token/quickstart'),
-                name: 'Confidential Transfer Fee Amount',
-                status: 'active',
-            };
-            break;
-        }
-        case 'confidentialTransferMint': {
-            const description =
-                'Allow token holders to opt-in to encrypted balances that are accessible only to them and the auditor';
-            return {
-                description,
-                externalLinks: populateExternalLinks('https://spl.solana.com/confidential-token/quickstart'),
-                name: 'Confidential Transfer',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'interestBearingConfig': {
-            const description = 'Allows the token balance to be displayed with accumulated interest';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('interest-bearing-token')),
-                name: 'Interest Bearing Token Configuration',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'transferFeeConfig': {
-            const description =
-                'Allows a fee to be set aside on every transfer that can only be withdrawn by the Withdraw Authority';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('transfer-fee')),
-                name: 'Transfer Fee',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'tokenGroup': {
-            return {
-                externalLinks: [],
-                name: 'Token Group',
-                status: 'active',
-            };
-            break;
-        }
-        case 'tokenGroupMember': {
-            return {
-                externalLinks: [],
-                name: 'Token Group Member',
-                status: 'active',
-            };
-            break;
-        }
-        case 'tokenMetadata': {
-            const description = 'Allows metadata to be written directly to the mint account';
-            return {
-                description,
-                externalLinks: populateExternalLinks(populateSolanaDevelopersLink('metadata-pointer')),
-                name: 'Token Metadata',
-                status: 'active',
-                tooltip: description,
-            };
-            break;
-        }
-        case 'unparseableExtension':
-        default:
-            return {
-                externalLinks: [],
-                name: 'Unknown Extension',
-                status: 'active',
-            };
-            break;
-    }
 }
